requirements.txt
numpy
pandas
matplotlib
scikit-learn
xarray
pyyaml
scikit-image
tifffile
zarr
readimc